[gui] improve custom widgets and ControlPanel
