+++ src/Gui/filelist.cmake
@@ -7,7 +7,8 @@
 set(gui_sources
     AboutDialog/AboutDialog.cpp
     BaseApplication.cpp
-    MaterialParameterEditor/MaterialParameterEditor.cpp
+    ParameterSetEditor/MaterialParameterEditor.cpp
+    ParameterSetEditor/ParameterSetEditor.cpp
     RadiumWindow/SimpleWindow.cpp
     SelectionManager/SelectionManager.cpp
     SkeletonBasedAnimation/SkeletonBasedAnimationUI.cpp
@@ -36,17 +37,15 @@ set(gui_sources
     Viewer/Viewer.cpp
     Viewer/WindowQt.cpp
     Widgets/ControlPanel.cpp
-    Widgets/PowerSlider.cpp
-    Widgets/CheckingSpinBox.cpp
     Widgets/MatrixEditor.cpp
-    Widgets/VectorEditor.cpp
 )
 
 set(gui_headers
     AboutDialog/AboutDialog.hpp
     BaseApplication.hpp
     MainWindowInterface.hpp
-    MaterialParameterEditor/MaterialParameterEditor.hpp
+    ParameterSetEditor/MaterialParameterEditor.hpp
+    ParameterSetEditor/ParameterSetEditor.hpp
     RaGui.hpp
     RadiumWindow/SimpleWindow.hpp
     RadiumWindow/SimpleWindowFactory.hpp
@@ -79,13 +78,14 @@ set(gui_headers
     Viewer/Viewer.hpp
     Viewer/WindowQt.hpp
     Widgets/ControlPanel.hpp
-    Widgets/PowerSlider.hpp
-    Widgets/CheckingSpinBox.hpp
     Widgets/MatrixEditor.hpp
+    Widgets/QtTypeWrapper.hpp
     Widgets/VectorEditor.hpp
 )
 
-set(gui_inlines Viewer/WindowQt.inl Widgets/ConstrainedNumericSpinBox.inl Widgets/ControlPanel.inl)
+set(gui_inlines Viewer/WindowQt.inl Widgets/ConstrainedNumericSpinBox.inl Widgets/ControlPanel.inl
+                Widgets/VectorEditor.inl
+)
 
 set(gui_uis AboutDialog/AboutDialog.ui SkeletonBasedAnimation/SkeletonBasedAnimationUI.ui
             Timeline/HelpDialog.ui Timeline/Timeline.ui
